Backport Missing `:via` Param
The `:via` param was not included on the gift card template, causing a
test in base to fail and subsequently block other plugin development
that depends on this gem. This change is from a later version of
Workarea, but was not backported to earlier versions. Additionally, this
commit adds the test workflow and ensures the correct version of
Workarea is being loaded to ensure accurate testing.

GIFTCARDS-7